# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ControlCreate": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/RiskCategory"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "kind": {
            "$ref": "#/components/schemas/ControlKind",
            "default": "preventivo"
          },
          "owner_membership_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Membership Id"
          },
          "review_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Date"
          },
          "risk_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Id"
          },
          "status": {
            "$ref": "#/components/schemas/ControlStatus",
            "default": "planejado"
          },
          "title": {
            "maxLength": 200,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "category"
        ],
        "title": "ControlCreate",
        "type": "object"
      },
      "ControlGraphOut": {
        "description": "Everything connected to one control: the Control Graph neighbourhood (D27).",
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionOut"
            },
            "title": "Actions",
            "type": "array"
          },
          "control": {
            "$ref": "#/components/schemas/ControlOut"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskRefOut"
            },
            "title": "Risks",
            "type": "array"
          }
        },
        "required": [
          "control",
          "risks",
          "actions",
          "evidence",
          "evidence_count"
        ],
        "title": "ControlGraphOut",
        "type": "object"
      },
...
      "ControlLinkOut": {
        "properties": {
          "control_id": {
            "format": "uuid",
            "title": "Control Id",
            "type": "string"
          },
          "linked": {
            "title": "Linked",
            "type": "boolean"
          },
          "risk_id": {
            "format": "uuid",
            "title": "Risk Id",
            "type": "string"
          }
        },
        "required": [
          "linked",
          "control_id",
          "risk_id"
        ],
        "title": "ControlLinkOut",
        "type": "object"
      },
      "ControlOut": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/RiskCategory"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentLinkOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ControlKind"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Date"
          },
          "status": {
            "$ref": "#/components/schemas/ControlStatus"
          },
          "template_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Code"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "category",
          "kind",
          "status",
          "template_code",
          "review_date",
          "created_at",
          "updated_at"
        ],
        "title": "ControlOut",
        "type": "object"
      },
      "ControlStatus": {
        "description": "Maturity ladder. `verificado` requires evidence linked to the control (service rule).",
        "enum": [
          "planejado",
          "parcial",
          "implementado",
          "verificado",
          "inativo"
        ],
        "title": "ControlStatus",
        "type": "string"
      },
      "ControlUpdate": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ControlKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner_membership_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Membership Id"
          },
          "review_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Date"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ControlStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ControlUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageOut": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "Page_ControlOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ControlOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ControlOut]",
        "type": "object"
      },
...
      "RiskCategory": {
        "enum": [
          "dados",
          "acesso",
          "seguranca",
          "fornecedores",
          "documentacao",
          "titulares",
          "incidentes",
          "pessoas"
        ],
        "title": "RiskCategory",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Compliance OS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/orgs/{org_id}/controls": {
      "get": {
        "operationId": "list_controls_api_v1_orgs__org_id__controls_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ControlStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/RiskCategory"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "owner_membership_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Membership Id"
            }
          },
          {
            "in": "query",
            "name": "risk_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ControlOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Controls",
        "tags": [
          "controls"
        ]
      },
      "post": {
        "operationId": "create_control_api_v1_orgs__org_id__controls_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Control",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/controls/{control_id}": {
      "delete": {
        "operationId": "delete_control_api_v1_orgs__org_id__controls__control_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Control Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Control",
        "tags": [
          "controls"
        ]
      },
      "get": {
        "operationId": "get_control_api_v1_orgs__org_id__controls__control_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Control Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlGraphOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Control",
        "tags": [
          "controls"
        ]
      },
      "patch": {
        "operationId": "update_control_api_v1_orgs__org_id__controls__control_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Control Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Control",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/controls/{control_id}/risks/{risk_id}": {
      "delete": {
        "operationId": "unlink_control_risk_api_v1_orgs__org_id__controls__control_id__risks__risk_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Control Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlLinkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unlink Control Risk",
        "tags": [
          "controls"
        ]
      },
      "post": {
        "operationId": "link_control_risk_api_v1_orgs__org_id__controls__control_id__risks__risk_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Control Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlLinkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link Control Risk",
        "tags": [
          "controls"
        ]
      }
    },
...
    "/api/v1/orgs/{org_id}/risks/{risk_id}/controls": {
      "get": {
        "operationId": "list_risk_controls_api_v1_orgs__org_id__risks__risk_id__controls_get",
        "parameters": [
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ControlOut"
                  },
                  "title": "Response List Risk Controls Api V1 Orgs  Org Id  Risks  Risk Id  Controls Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Risk Controls",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/risks/{risk_id}/plan": {
      "post": {
        "operationId": "plan_risk_api_v1_orgs__org_id__risks__risk_id__plan_post",
        "parameters": [
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan Risk",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/risks/{risk_id}/recommendation": {
      "get": {
        "operationId": "risk_recommendation_api_v1_orgs__org_id__risks__risk_id__recommendation_get",
        "parameters": [
          {
            "in": "path",
            "name": "risk_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Risk Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk Recommendation",
        "tags": [
          "controls"
        ]
      }
    },
...
  }
}